+++ src/hal/actor/commands/status.py
@@ -8,11 +8,17 @@
 
 from __future__ import annotations
 
+from typing import TYPE_CHECKING
+
 import click
 
 from clu import Command
 
-from . import HALCommandType, hal_command_parser
+from . import hal_command_parser
+
+
+if TYPE_CHECKING:
+    from . import HALCommandType
 
 
 __all__ = ["status"]