+++ src/hal/helpers/__init__.py
@@ -61,7 +61,9 @@ class HALHelper:
 
 from .apogee import *
 from .boss import *
+from .cherno import *
 from .ffs import *
+from .jaeger import *
 from .lamps import *
 from .scripts import *
 from .tcc import *